update controllers and add middleware to respective files
