+++ middleware/authMiddleware.js
@@ -1,0 +1,8 @@
+const { expressjwt } = require("express-jwt");
+const config = require("../config");
+
+
+exports.validateToken = expressjwt({
+    secret: config.jwtSecret,
+    algorithms: ["HS256"],
+});
